Add files via upload
The code above fetches the job page and then parses the HTML code to find the requirements section. The program then creates a list of keywords that match the requirements, and then creates a dictionary that counts the number of occurrences of each keyword. At the end, the program prints out the keywords and their number of occurrences.